+++ driver/src/test/unit/com/mongodb/rx/client/MongoDatabaseSpecification.groovy
@@ -43,7 +43,7 @@ class MongoDatabaseSpecification extends Specification {
     def 'should have the same methods as the wrapped MongoDatabase'() {
         given:
         def exclusions = ['getObservableAdapter', 'withObservableAdapter']
-        def wrapped = WrappedMongoDatabase.methods*.name.sort()
+        def wrapped = WrappedMongoDatabase.methods*.name.sort() - 'createView'
         def local = MongoDatabase.methods*.name.sort() - exclusions
 
         expect: