+++ docs/reference/content/changelog.md
@@ -11,6 +11,15 @@ title = "Changelog"
 Changes between released versions
 
 
+### 1.3
+
+
+[Full change list](https://jira.mongodb.org/issues/?jql=project%20%3D%20JAVARX%20AND%20fixVersion%20%3D%201.3)
+
+  * Updated MongoDB Driver Async to 3.4.0-beta1
+    
+  * Updated RxJava to 1.2.0 [JAVARX-33](https://jira.mongodb.org/browse/JAVARX-33)
+
 ### 1.2 
 
 [Full change list](https://jira.mongodb.org/issues/?jql=project%20%3D%20JAVARX%20AND%20fixVersion%20%3D%201.2)